+CORS và is_locked

+++ dtb.py
@@ -22,6 +22,7 @@ class NoEncryptedCustomer(Model):
     Balance = IntegerField()
     Exp_date_card = TextField()
     Types_of_ticket = IntegerField()
+    Is_locked = BooleanField()
 
     class Meta:
         table_name = 'No_Encrypted_Customer'
@@ -42,7 +43,8 @@ def select_user(id):
         'Exp_date_ticket': customer['Exp_date_ticket'],
         'Balance': customer['Balance'],
         'Exp_date_card': customer['Exp_date_card'],
-        'Types_of_ticket': customer['Types_of_ticket']
+        'Types_of_ticket': customer['Types_of_ticket'],
+        'Is_locked': customer['Is_locked']
     }
 
 
@@ -70,21 +72,44 @@ def add_user(name, cmnd, gender, picture, dob, exp_date_ticket, balance, exp_dat
 
 def update_user(id, name, cmnd, gender, picture, dob, exp_date_ticket, balance, exp_date_card, types_of_ticket):
     record = NoEncryptedCustomer.get_by_id(id)
-    record.Name = name
-    record.CMND = cmnd
-    record.Gender = gender
-    record.Picture = picture
-    record.DoB = dob
-    record.Exp_date_ticket = exp_date_ticket
-    record.Balance = balance
-    record.Exp_date_card = exp_date_card
-    record.Types_of_ticket = types_of_ticket
-    record.save()
+    if not record.Is_locked:
+        record.Name = name
+        record.CMND = cmnd
+        record.Gender = gender
+        record.Picture = picture
+        record.DoB = dob
+        record.Exp_date_ticket = exp_date_ticket
+        record.Balance = balance
+        record.Exp_date_card = exp_date_card
+        record.Types_of_ticket = types_of_ticket
+        record.save()
+    else:
+        raise Exception('Locked user')
 
 
 def delete_user(id):
     record = NoEncryptedCustomer.get_by_id(id)
-    record.delete_instance()
+    if not record.Is_locked:
+        record.delete_instance()
+    else:
+        raise Exception('Locked user')
+
+
+def lock_user(id):
+    record = NoEncryptedCustomer.get_by_id(id)
+    if not record.Is_locked:
+        record.Is_locked = True
+        record.save()
+    else:
+        raise Exception('User is already locked')
+
+
+def unlock_user(id):
+    record = NoEncryptedCustomer.get_by_id(id)
+    if not record.Is_locked:
+        raise Exception('User is not locked')
+    else:
+        record.Is_locked = False
 
 
 if __name__ == '__main__':
